requirements.txt
alembic==1.4.2
aniso8601==8.0.0
argcomplete==1.12.0
bcrypt==3.1.7
boto3==1.14.21
botocore==1.17.21
certifi==2020.6.20
cffi==1.14.0
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.1
docutils==0.15.2
durationpy==0.5
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
flask-marshmallow==0.11.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
future==0.18.2
gunicorn==20.0.4
hjson==3.0.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.10.0
kappa==0.6.0
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.22.3
pip-tools==5.2.1
pkg-resources==0.0.0
placebo==0.9.0
psycopg2-binary==2.8.4
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.6.1
python-editor==1.0.4
python-slugify==4.0.1
pytz==2019.3
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.3
six==1.14.0
SQLAlchemy==1.3.15
text-unidecode==1.3
toml==0.10.1
tqdm==4.47.0
troposphere==2.6.2
urllib3==1.25.9
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.51.0

python-dotenv~=0.14.0